added another hook for mock rides
