+++ apps/ui/src/hooks/Communities/useGetAllCommunities.tsx
@@ -1,8 +1,10 @@
-import { ClientCommunity } from '../../types/community-type';
-import { getRandomOption } from '../../utils';
+import { Community } from '@communecar/types';
+
 import tlv from '../../assets/tlv.png';
 import apple from '../../assets/apple.png';
 import camera from '../../assets/camera.png';
+
+import { getRandomOption } from '../../utils';
 
 const options = [tlv, apple, camera];
 
@@ -14,7 +16,7 @@ const picturesUrl = [
   getRandomOption(options),
 ];
 
-const communities: ClientCommunity[] = [
+const communities: Community[] = [
   {
     name: 'Travel friends Haifa - Tel Aviv',
     description:
@@ -40,8 +42,47 @@ const communities: ClientCommunity[] = [
   },
 ];
 
+const rides = [
+  {
+    driver: 'Zoe Shwartz',
+    departureTime: new Date(Date.now() + 60 * 60000), // Adding 60 minutes to the current time
+    communityName: 'Travel friends Haifa - Tel Aviv',
+    startLocation: 'Rotchild street, Tel Aviv',
+    png: getRandomOption(options),
+    destination: 'Pardesia',
+  },
+  {
+    driver: 'Dar Nachmani',
+    departureTime: new Date(Date.now() + 120 * 60000),
+    communityName: 'Apple Friends - IL',
+    png: getRandomOption(options),
+    startLocation: 'Efraim Katzir street, Hod Hasharon',
+    destination: 'Modiin',
+  },
+  {
+    driver: 'Avi Ron',
+    departureTime: new Date(Date.now() + 50 * 60000),
+    communityName: 'Travel friends Haifa - Tel Aviv',
+    startLocation: 'Weizman street, Petah Tikva',
+    destination: 'Holon',
+    png: getRandomOption(options),
+  },
+  {
+    driver: 'Tal Kovler',
+    departureTime: new Date(Date.now() + 25 * 60000),
+    communityName: 'Apple Friends - IL',
+    startLocation: 'Bla street, Haifa',
+    destination: 'The Golan',
+    png: getRandomOption(options),
+  },
+];
+
 const useGetAllCommunities = () => {
   return communities;
 };
 
-export { useGetAllCommunities };
+const useGetAllRides = () => {
+  return rides;
+};
+
+export { useGetAllCommunities, useGetAllRides };
